with add test

+++ src/main/java/gene/chunk/Chunk.java
@@ -4,7 +4,7 @@ public interface Chunk {
 
     public double getHash();
 
-    public long[] getBits();
+    public int[] getBits();
 
     public double getChunkSum();
